Updates to the NCO mode (#443)
* Add a move step to binaries for NCO purposes.

* Pass COMROOT to setpdy.

* Correct path of post dependency on forecat.

* Remove DATA_SHARED directory.

* Remove symlinks and use unique id for run_fcst instead.

* Remove forecast directory after post is done.

* Use raw workflow_id

* Add templated data paths to wcoss2 that use compath.py.

* Remove cron relaunch from config.nco

* Undo for_ICS/LBCS stuff in NCO mode.

* Remove RESTART symlinking because DATAFCST is accessible to all other tasks.

* Undo cyc addition to COMIN.

* Update devclean.

* Follow symlinks when tarring log files in Jenkins.

* Add memory otpion for get_extrn_ics/lbcs.

* Optimize fundamental test coverage.

* Reduce work load on orion.

* Add minimum memory requirement to all serial tasks.

* Simplify set_predef_grid_params.

* Simplify generate_FV3LAM_wflow.

* Simplify link_fix.

* Simplify to get rid of export_vars from setup.

* Import only necessary variables from environment.

* Use `eval echo` or `eval` depending on situation.

* Add cyc back to COMIN.

* Fix calculate_cost.

* Fix gaea test list.

* Remove a test case from cheyenne that is taking extremely long.

* Reduce memory requirement for serial job to 1G.

* Create COMINaws ics/lbcs staging directory.

* Also symlink gfs ics/lbcs if on disk and in community mode.

* Remove failing tests on orion.

* Exclude gaea from --mem specification.

* Add a run_vx modulefile for orion.

* Add separate fundamental list for cheyenne GNU runs.

* Bug fix for wcoss2 GFS version

* Load prod_util on wcoss2.

* Rename AWSROOT to EXTROOT.

* Remove leftover COMINaws.

+++ devbuild.sh
@@ -33,6 +33,8 @@ OPTIONS
       does a "make clean"
   --build
       does a "make" (build only)
+  --move
+      move binaries to final location.
   --build-dir=BUILD_DIR
       build directory
   --install-dir=INSTALL_DIR
@@ -125,6 +127,7 @@ BUILD_RRFS_UTILS="off"
 # Make options
 CLEAN=false
 BUILD=false
+MOVE=false
 USE_SUB_MODULES=false #change default to true later
 
 # process required arguments
@@ -155,6 +158,7 @@ while :; do
     --continue=?*|--continue=) usage_error "$1 argument ignored." ;;
     --clean) CLEAN=true ;;
     --build) BUILD=true ;;
+    --move) MOVE=true ;;
     --build-dir=?*) BUILD_DIR=${1#*=} ;;
     --build-dir|--build-dir=) usage_error "$1 requires argument." ;;
     --install-dir=?*) INSTALL_DIR=${1#*=} ;;
@@ -398,18 +402,31 @@ module list
 mkdir -p ${BUILD_DIR}
 cd ${BUILD_DIR}
 
-printf "... Generate CMAKE configuration ...\n"
-cmake ${SRW_DIR} ${CMAKE_SETTINGS} 2>&1 | tee log.cmake
-
 if [ "${CLEAN}" = true ]; then
-    printf "... Clean executables ...\n"
-    make ${MAKE_SETTINGS} clean 2>&1 | tee log.make
+    if [ -f $PWD/Makefile ]; then
+       printf "... Clean executables ...\n"
+       make ${MAKE_SETTINGS} clean 2>&1 | tee log.make
+    fi
 elif [ "${BUILD}" = true ]; then
+    printf "... Generate CMAKE configuration ...\n"
+    cmake ${SRW_DIR} ${CMAKE_SETTINGS} 2>&1 | tee log.cmake
+
     printf "... Compile executables ...\n"
     make ${MAKE_SETTINGS} build 2>&1 | tee log.make
 else
+    printf "... Generate CMAKE configuration ...\n"
+    cmake ${SRW_DIR} ${CMAKE_SETTINGS} 2>&1 | tee log.cmake
+
     printf "... Compile and install executables ...\n"
     make ${MAKE_SETTINGS} install 2>&1 | tee log.make
+
+    if [ "${MOVE}" = true ]; then
+       if [[ ! ${SRW_DIR} -ef ${INSTALL_DIR} ]]; then
+           printf "... Moving executables to final locations ...\n"
+           mkdir -p ${SRW_DIR}/${BIN_DIR}
+           mv ${INSTALL_DIR}/${BIN_DIR}/* ${SRW_DIR}/${BIN_DIR}
+       fi
+    fi
 fi
 
 exit 0
